+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Bug #73: Update attribute retrieval in tests to use `getAttribute()` method for consistency (@terabytesoftw)
 - Enh #74: Add `Line` class and corresponding tests for SVG `<line>` element functionality (@terabytesoftw)
 - Bug #75: Update documentation for `Circle`, `Ellipse`, and `Image` class to correct terminology from container element to element (@terabytesoftw)
+- Enh #76: Add `HasGradientTransform` trait and corresponding tests for managing SVG `gradientTransform` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024